requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9
pdfminer.six==20221105
python-docx==1.1.0
ebooklib==0.18
beautifulsoup4==4.12.3
pydub==0.25.1
aiofiles==23.2.1
httpx==0.27.0
lxml==5.2.1
soundfile>=0.12.1
numpy>=1.20.0,<2.0.0
